+++ src/v_video.h
@@ -245,7 +245,7 @@ public:
 	void STACK_ARGS DrawTextA (FFont *font, int normalcolor, int x, int y, const char *string, ...);
 #endif
 	void DrawTextV (FFont *font, int normalcolor, int x, int y, const char *string, va_list tags);
-	void STACK_ARGS DrawChar (FFont *font, int normalcolor, int x, int y, BYTE character, ...);
+	void STACK_ARGS DrawChar (FFont *font, int normalcolor, int x, int y, int character, ...);
 
 	struct DrawParms
 	{
